Implement basic strategie

+++ CrashGameLoadTest/Interfaces/IBetDecisionStrategy.cs
@@ -5,6 +5,6 @@ namespace CrashGameLoadTest.Interfaces
     public interface IBetDecisionStrategy
     {
         Task<bool> ShouldBetAsync(PlayerContext context, CancellationToken cancellationToken);
-        Task<decimal> GetBetAmountAsync(PlayerContext context, CancellationToken cancellationToken);
+        Task<double> GetBetAmountAsync(PlayerContext context, CancellationToken cancellationToken);
     }
 }
