+++ CrashGameLoadTest/Models/EventModels/InitialDataResult.cs
@@ -15,9 +15,9 @@ namespace CrashGameLoadTest.Models.EventModels
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string? Language { get; set; }
-        public decimal MinBet { get; set; }
-        public decimal MaxBet { get; set; }
-        public decimal MaxWin { get; set; }
+        public double MinBet { get; set; }
+        public double MaxBet { get; set; }
+        public double MaxWin { get; set; }
         public bool? IsDemo { get; set; }
         public bool IsMaintenance { get; set; }
         public int? MaintenanceStartDateInSeconds { get; set; }
@@ -33,7 +33,7 @@ namespace CrashGameLoadTest.Models.EventModels
         public double? ElapsedOddTime { get; set; }
         public DateTime? StartOddTime { get; set; }
         public double? CrashPoint { get; set; }
-        public decimal? BonusBankAmount { get; set; }
+        public double? BonusBankAmount { get; set; }
     }
 
     public class User
@@ -43,13 +43,13 @@ namespace CrashGameLoadTest.Models.EventModels
         public long DisplayId { get; set; }
         public string Currency { get; set; }
         public BonusBankWinData? BonusBankWinData { get; set; }
-        public decimal Balance { get; set; }
-        public decimal Rate { get; set; }
+        public double Balance { get; set; }
+        public double Rate { get; set; }
     }
 
     public class BonusBankWinData
     {
-        public decimal? BankWin { get; set; }
+        public double? BankWin { get; set; }
         public BetSectionEnum BetSection { get; set; }
     }
 }
